+++ tests/Model/BaseManagerTest.php
@@ -15,6 +15,8 @@ namespace Sonata\Doctrine\Tests\Model;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\Persistence\ManagerRegistry;
+use Doctrine\Persistence\ObjectManager;
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Sonata\Doctrine\Model\BaseManager;
 
@@ -41,13 +43,33 @@ class ManagerTest extends BaseManager
  */
 final class BaseManagerTest extends TestCase
 {
+    /**
+     * @var MockObject&ObjectManager
+     */
+    private $objectManager;
+
+    protected function setUp(): void
+    {
+        $this->objectManager = $this->createMock(ObjectManager::class);
+
+        $managerRegistry = $this->createStub(ManagerRegistry::class);
+        $managerRegistry->method('getManagerForClass')->willReturn($this->objectManager);
+
+        $this->manager = new ManagerTest('class', $managerRegistry);
+    }
+
     public function testCheckObject(): void
     {
         $this->expectException(\InvalidArgumentException::class);
         $this->expectExceptionMessage('Object must be instance of class, DateTime given');
 
-        $manager = new ManagerTest('class', $this->createMock(ManagerRegistry::class));
+        $this->manager->publicCheckObject(new \DateTime());
+    }
 
-        $manager->publicCheckObject(new \DateTime());
+    public function testClearManager()
+    {
+        $this->objectManager->expects($this->once())->method('clear');
+
+        $this->manager->clear();
     }
 }