+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.13.0](https://github.com/sonata-project/sonata-doctrine-extensions/compare/1.12.0...1.13.0) - 2021-06-15
+### Added
+- [[#323](https://github.com/sonata-project/sonata-doctrine-extensions/pull/323)] Added `ClearableManagerInterface` to be able to clear the manager in an easy way ([@jordisala1991](https://github.com/jordisala1991))
+
+### Fixed
+- [[#298](https://github.com/sonata-project/sonata-doctrine-extensions/pull/298)] Dependency on a non-existent service "sonata.doctrine.adapter.doctrine_phpcr" ([@skydiablo](https://github.com/skydiablo))
+
 ## [1.12.0](https://github.com/sonata-project/sonata-doctrine-extensions/compare/1.11.0...1.12.0) - 2021-03-11
 ### Added
 - [[#303](https://github.com/sonata-project/sonata-doctrine-extensions/pull/303)] Add type hints to interfaces ([@core23](https://github.com/core23))